fix : no ssr client component to use 'dynamic'
