+++ src/components/imageUploader/index.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { memo } from 'react';
 import dynamic from 'next/dynamic';
